chore(src/FMMCompute.cpp): int -> size_t in index

+++ src/FMMCompute.cpp
@@ -17,7 +17,7 @@ void convert_to_numpy(const vector<double> & input, object&  obj)
     double *p = (double*)buf;
     Py_XDECREF(pobj);
 
-    for (int i  = 0; i < input.size(); i++)
+    for (size_t i  = 0; i < input.size(); i++)
     {
         p[i] = input[i];
     }
